petites modifs de rien du tout

+++ start.bash
@@ -15,14 +15,6 @@ package_name="home_package.tar.xz"
 package_location="$HOME/.cache/$package_name"
 package_version_location="$HOME/.cache/home_package_version"
 
-if [ -f "$package_name" ]; then
-  package_location="$(realpath "$package_name")"
-  echo "install: $package_location"
-  install_package "$package_location"
-  echo "$package_location" > "$package_version_location"
-  exit
-fi
-
 package_installed_version="$(cat "$HOME/.cache/home_package_version")"
 if [ "$package_installed_version" ]; then
   echo "package_installed_version: $package_installed_version"
@@ -37,15 +29,22 @@ else
   echo "package_latest_version: not found (https://github.com/L0L022/config_iut)"
 fi
 
-if [[ "$package_latest_version" && "$package_installed_version" != "$package_latest_version" ]]; then
-  echo "install: $package_latest_version"
-  echo "download"
-  if [ -f "$package_location" ]; then
-    rm "$package_location"
-  fi
-  curl -sL -o "$package_location" https://github.com/L0L022/config_iut/releases/download/"$package_latest_version"/"$package_name"
+if [ -f "$package_name" ]; then
+  package_location="$(realpath "$package_name")"
+  echo "install: $package_location"
   install_package "$package_location"
-  echo "$package_latest_version" > "$package_version_location"
+  echo "$package_location" > "$package_version_location"
+else
+  if [[ "$package_latest_version" && "$package_installed_version" != "$package_latest_version" ]]; then
+    echo "install: $package_latest_version"
+    echo "download"
+    if [ -f "$package_location" ]; then
+      rm "$package_location"
+    fi
+    curl -sL -o "$package_location" https://github.com/L0L022/config_iut/releases/download/"$package_latest_version"/"$package_name"
+    install_package "$package_location"
+    echo "$package_latest_version" > "$package_version_location"
+  fi
 fi
 
 echo "install crazy_patch.bash"
